Move deprecatedSnapshot resolvers into /resolvers/snapshots.js
